+++ src/game/assets.rs
@@ -2,6 +2,7 @@ extern crate find_folder;
 extern crate opengl_graphics;
 
 use std::path::PathBuf;
+use graphics::ImageSize;
 use opengl_graphics::{ TextureSettings, Texture};
 
 pub struct Assets {}
@@ -25,5 +26,41 @@ impl Assets {
 
     fn from_path(path: &str) -> Texture {
         return Texture::from_path(Assets::assets(path), &TextureSettings::new()).unwrap();
+    }
+}
+
+pub struct Background {
+    pub levels: [Texture; 4],
+    pub translations: [f64; 4],
+}
+
+impl Background {
+    pub fn new() -> Background {
+        Background {
+            levels: [
+                Assets::texture("parallax-forest-back-trees.png"),
+                Assets::texture("parallax-forest-middle-trees.png"),
+                Assets::texture("parallax-forest-lights.png"),
+                Assets::texture("parallax-forest-front-trees.png"),
+            ],
+            translations: [0.0, 0.0, 0.0, 0.0],
+        }
+    }
+
+    pub fn animate(&mut self) {
+        self.translations[0] -= 0.03;
+        self.translations[1] -= 0.06;
+        self.translations[3] -= 0.2;
+
+        self.reset();
+    }
+
+    fn reset(&mut self) {
+        let min: f64 = -1.0 * self.levels[0].get_width() as f64;
+        for i in 0..4 {
+            if self.translations[i] < min {
+                self.translations[i] = 0.0;
+            }
+        }
     }
 }
