+++ src/game/mod.rs
@@ -1,5 +1,6 @@
 pub use self::canvas::Canvas;
 pub use self::assets::Assets;
+pub use self::assets::Background;
 
 mod canvas;
 mod assets;
